upgrade and fix comments

+++ sys/inc/includes/delete_comment.php
@@ -5,18 +5,23 @@ $id = (!empty($id)) ? (int)$id : 0;
 if ($id < 1) redirect('/' . $this->module);
 
 
-$commClassName = ucfirst($this->module) . 'CommentsModel';
-$commModel = new $commClassName;
-$comment = $commModel->getById($id);
-if ($comment) {
-	$entityID = $comment->getEntity_id();
-	$comment->delete();
-	
-	$entity = $this->Model->getById($entityID);
-	$entity->setComments($entity->getComments() - 1);
-	$entity->save();
+$commentsModel = $this->Register['ModManager']->getModelInstance($this->module . 'Comments');
+if ($commentsModel) {
+	$comment = $commentsModel->getById($id);
+	if ($comment) {
+		$entityID = $comment->getEntity_id();
+		$comment->delete();
+		
+		$entity = $this->Model->getById($entityID);
+		if ($entity) {
+			$entity->setComments($entity->getComments() - 1);
+			$entity->save();
+			
+
+
+			if ($this->Log) $this->Log->write('delete comment for ' . $this->module, $this->module . ' id(' . $entityID . ')');
+			return $this->showInfoMessage(__('Comments is deleted'), '/' . $this->module . '/view/' . $entityID );
+		}
+	}
 }
-
-
-if ($this->Log) $this->Log->write('delete comment for ' . $this->module, $this->module . ' id(' . $entityID . ')');
-return $this->showInfoMessage(__('Comments is deleted'), '/' . $this->module . '/view/' . $entityID );
+return $this->showInfoMessage(__('Some error occurred'), '/' . $this->module . '/view/' . $entityID);
